requirements.txt
docarray[hnswlib]
docx2txt
faiss-gpu
instructorembedding
langchain
pdfplumber
python-dotenv
sentence-transformers
torch
tqdm
transformers
unstructured[local-inference]
